backend/tools/gemini_client.py
"""
News Validator Agent - Gemini API Client
Handles integration with Google Gemini API for claim extraction and analysis
"""

import os
from typing import List, Dict, Any, Optional
import google.generativeai as genai


class GeminiClient:
    """
    Client for interacting with Google Gemini API
    """
    
    def __init__(self, api_key: str = None):
        """
        Initialize Gemini client
        
        Args:
            api_key: Google Gemini API key (if None, reads from environment)
        """
        self.api_key = api_key or os.getenv('GEMINI_API_KEY')
        if not self.api_key:
            raise ValueError("Gemini API key is required. Set GEMINI_API_KEY environment variable.")
        
        # TODO: Configure Gemini client
        # genai.configure(api_key=self.api_key)
        # self.model = genai.GenerativeModel('gemini-pro')
        
    async def extract_claims(self, text: str) -> List[Dict[str, Any]]:
        """
        Extract key claims from news text
        
        Args:
            text: News article or content text
            
        Returns:
            List of extracted claims with metadata
        """
        prompt = f"""
        Analyze the following news content and extract key factual claims.
        For each claim, provide:
        1. The claim statement
        2. Confidence level (0-1)
        3. Category (political, economic, scientific, etc.)
        4. Verifiability (easy/medium/hard to verify)
        
        News content: {text}
        
        Return as structured JSON.
        """
        
        # TODO: Implement actual Gemini API call
        # response = await self.model.generate_content_async(prompt)
        # return self._parse_claims_response(response.text)
        
        # Placeholder response
        return [
            {
                "claim": "Sample extracted claim",
                "confidence": 0.8,
                "category": "general",
                "verifiability": "medium"
            }
        ]
    
    async def analyze_credibility(self, claim: str, sources: List[str]) -> Dict[str, Any]:
        """
        Analyze credibility of a claim against multiple sources
        
        Args:
            claim: The claim to analyze
            sources: List of source texts to compare against
            
        Returns:
            Credibility analysis with score and reasoning
        """
        prompt = f"""
        Analyze the credibility of this claim against the provided sources:
        
        Claim: {claim}
        
        Sources:
        {chr(10).join([f"Source {i+1}: {source}" for i, source in enumerate(sources)])}
        
        Provide:
        1. Overall credibility score (0-1)
        2. Supporting evidence count
        3. Contradicting evidence count
        4. Key contradictions found
        5. Reasoning for the score
        
        Return as structured JSON.
        """
        
        # TODO: Implement actual Gemini API call
        # response = await self.model.generate_content_async(prompt)
        # return self._parse_credibility_response(response.text)
        
        # Placeholder response
        return {
            "credibility_score": 0.7,
            "supporting_evidence": 2,
            "contradicting_evidence": 1,
            "contradictions": [],
            "reasoning": "Placeholder credibility analysis"
        }
    
    async def detect_contradictions(self, sources: List[str]) -> List[Dict[str, Any]]:
        """
        Detect contradictions between multiple news sources
        
        Args:
            sources: List of source texts to compare
            
        Returns:
            List of detected contradictions with details
        """
        prompt = f"""
        Compare these news sources and identify any contradictions:
        
        {chr(10).join([f"Source {i+1}: {source}" for i, source in enumerate(sources)])}
        
        For each contradiction found, provide:
        1. Contradicting statements
        2. Source indices involved
        3. Severity (minor/major)
        4. Topic area
        
        Return as structured JSON.
        """
        
        # TODO: Implement actual Gemini API call
        # response = await self.model.generate_content_async(prompt)
        # return self._parse_contradictions_response(response.text)
        
        # Placeholder response
        return [
            {
                "statement_1": "Sample contradicting statement 1",
                "statement_2": "Sample contradicting statement 2",
                "sources": [0, 1],
                "severity": "minor",
                "topic": "general"
            }
        ]
    
    def _parse_claims_response(self, response_text: str) -> List[Dict[str, Any]]:
        """Parse Gemini response for claim extraction"""
        # TODO: Implement response parsing
        return []
    
    def _parse_credibility_response(self, response_text: str) -> Dict[str, Any]:
        """Parse Gemini response for credibility analysis"""
        # TODO: Implement response parsing
        return {}
    
    def _parse_contradictions_response(self, response_text: str) -> List[Dict[str, Any]]:
        """Parse Gemini response for contradiction detection"""
        # TODO: Implement response parsing
        return []
