backend/requirements.txt
# News Validator Agent - Backend Dependencies

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client for API calls
aiohttp==3.9.1
httpx==0.25.2

# Google Gemini API
google-generativeai==0.3.2

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# PostgreSQL
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0

# Date/time handling
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code quality (development)
black==23.11.0
flake8==6.1.0
mypy==1.7.1
